+++ site/index.js
@@ -75,7 +75,7 @@ class Site extends ShallowComponent {
                             <NavItem eventKey="Recharts" className="re-charts">
                                 {"<Recharts />"}
                             </NavItem>
-                            <NavDropdown title="Language" id="nav-dropdown" style={{ display: "none" }}>
+                            <NavDropdown title="Language" id="nav-dropdown" style={{ display: "inline" }}>
                                 <MenuItem eventKey="en_US">English</MenuItem>
                                 <MenuItem eventKey="tr_TR">Türkçe</MenuItem>
                             </NavDropdown>